[issue-1] save curent progress
