+++ src/lib/queue.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <sys/types.h>
 
-#include "queue.h"
+#include "cloudtiering.h"
 
 
 /**
