--- src/tst/test-app.h
@@ -1,7 +1,0 @@
-#ifndef TEST_APP_H
-#define TEST_APP_H
-
-/* add include statements when any test implemented for app */
-
-#endif // TEST_APP_H
-
