--- src/tst/test-lib.h
@@ -1,8 +1,0 @@
-#ifndef TEST_LIB_H
-#define TEST_LIB_H
-
-#include "../lib/conf.h"
-#include "../lib/queue.h"
-
-#endif // TEST_LIB_H
-
